+++ tools/shot.gd
@@ -11,6 +11,8 @@ var _n := 0
 var _moved := false
 var _flagdemo := false
 var _flagdrop := false
+var _scene := "res://scenes/main.tscn"
+var _vis := ""   # --show=NodeVar: make current_scene.<var> visible, hide _menu_root
 
 
 func _initialize() -> void:
@@ -23,10 +25,14 @@ func _initialize() -> void:
 			_flagdemo = true
 		elif a == "--flagdrop":
 			_flagdrop = true
+		elif a.begins_with("--scene="):
+			_scene = a.substr(8)
+		elif a.begins_with("--show="):
+			_vis = a.substr(7)
 		elif a.begins_with("--at="):
 			var xy := a.substr(5).split(",")
 			_at = Vector2(float(xy[0]), float(xy[1]))
-	change_scene_to_file("res://scenes/main.tscn")
+	change_scene_to_file(_scene)
 
 
 func _process(_delta: float) -> bool:
@@ -46,6 +52,13 @@ func _process(_delta: float) -> bool:
 			if _flagdrop and fl.size() >= 2:
 				# RED flag lying in the field next to the player.
 				fl[1].position = current_scene._settle_on_ground(_at + Vector2(90, -20))
+	if _vis != "" and _n == 10 and current_scene != null:
+		var mr = current_scene.get("_menu_root")
+		if mr != null:
+			mr.visible = false
+		var panel = current_scene.get(_vis)
+		if panel != null:
+			panel.visible = true
 	if _n == _frames:
 		var img := root.get_texture().get_image()
 		img.save_png(_out)